Configuration / Embedding (#13)
* Remove extraneous HTML

* Isolate JS selectors from CSS selectors

* USe a class selector for consistency

* Construct DOM for the REPL

One objective is to have the REPL bootstrap itself by creating it's own
DOM and binding behaviour to it (rather than presuming some existing DOM
will be present).

In this commit, a very basic little DOM construction library is
introduced. It's purpose is to take a simple JavaScript data structure
representing the DOM, build the appropriate DOM and mutate the structure
with references to the constructed DOM - and give it back. This should
provide a way to access DOM without using the expensive DOM API - and
modify our DOM in a declaritive fashion, should we wish.

Additionally, some minimal structure representing the REPL is declared,
and a set of lenses that will provide access to interesting parts of the
DOM when applied to the structure.

* CSS classes are only needed for styling purposes

* Remove HTML

* Correct some CSS

* Update CSS selector names

* Bind behaviour after entering the DOM

* CodeMirror is not interactive if the textarea it replaces is hidden

* Space out the short-url button

* Demonstrate embedding

* REPL can take inital value from a config

* Select elements with querySelector, remove lenses.

Using the lenses was a bit brittle - changing the DOM would also mean
changing the lens. Considered dynamically generating the lens and
decided the added complexity was not a great trade for simply using
querySelector.

* Remove queryString

* Use config to update the queryString

* Use config to set the api / return urls

* Hide url controls if the api configuration is incomplete

* Update bundle

* Install rdom

* Replace DOM builder with rdom@0.6.0

* Update bundle

* Remove wzrd.in dependencies

* Remove extraneous file, declare 'pre' tag in rdom